src/jetson/server/requirements.txt
Flask==2.0.3
websocket-client==1.3.1
rel==0.4.9